requirements.txt
pandas == 2.1.1
numpy == 1.26.4
python-dateutil == 2.9.0
pytz == 2024.1
six == 1.16.0
pyyaml